# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: ArtiCurated Order Management API
  description: |
    RESTful API for managing orders and returns in the ArtiCurated boutique marketplace.

    ## Key Features
    - Deterministic state machine for order lifecycle management
    - Multi-step returns workflow with manager approval
    - Comprehensive audit trail for all state transitions
    - Asynchronous background processing for invoices and refunds

    ## State Machines

    ### Order States
    - `PENDING_PAYMENT` → `PAID` → `PROCESSING_IN_WAREHOUSE` → `SHIPPED` → `DELIVERED`
    - `PENDING_PAYMENT` or `PAID` can transition to `CANCELLED`

    ### Return States
    - `REQUESTED` → `APPROVED` → `IN_TRANSIT` → `RECEIVED` → `COMPLETED`
    - `REQUESTED` can transition to `REJECTED` (terminal)

  version: 1.0.0
  contact:
    name: ArtiCurated Engineering
    email: engineering@articurated.com
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.articurated.com/v1
    description: Production server
...
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Create new order
      description: Creates a new order with initial state PENDING_PAYMENT
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
            example:
              customer_id: "550e8400-e29b-41d4-a716-446655440000"
              line_items:
                - product_id: "650e8400-e29b-41d4-a716-446655440001"
                  product_name: "Handcrafted Ceramic Vase"
                  quantity: 1
                  unit_price: 129.99
              shipping_address:
                street: "123 Main St"
                city: "San Francisco"
                state: "CA"
                zip: "94102"
                country: "USA"
              billing_address:
                street: "123 Main St"
                city: "San Francisco"
                state: "CA"
                zip: "94102"
                country: "USA"
              payment_method: "credit_card"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"

    get:
      tags: [Orders]
      summary: List orders
      description: Retrieve paginated list of orders with optional filtering
      operationId: listOrders
      parameters:
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrderResponse"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order details
      description: Retrieve complete order information including line items
      operationId: getOrder
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{order_id}/state:
    patch:
      tags: [Orders]
      summary: Transition order state
      description: |
        Update order to new state with validation against allowed transitions.
        Records audit trail entry for compliance.

        **Triggers:**
        - SHIPPED → Triggers invoice generation background job
      operationId: updateOrderState
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StateTransition"
            example:
              new_state: "PAID"
              reason: "Payment confirmed via Stripe"
      responses:
        "200":
          description: State transition successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "INVALID_STATE_TRANSITION"
                  message: "Cannot transition from SHIPPED to PAID"
                  details:
                    current_state: "SHIPPED"
                    requested_state: "PAID"
                    allowed_transitions: ["DELIVERED"]
...
  /orders/{order_id}/history:
    get:
      tags: [Orders]
      summary: Get order state history
      description: Retrieve complete audit trail of order state transitions
      operationId: getOrderHistory
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order state history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StateHistoryRecord"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== RETURN ENDPOINTS ====================
...
components:
  parameters:
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique order identifier
...
  schemas:
...
    OrderStatus:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAID
        - PROCESSING_IN_WAREHOUSE
        - SHIPPED
        - DELIVERED
        - CANCELLED
      description: Order lifecycle states
...
    ActorType:
      type: string
      enum:
        - USER
        - SYSTEM
      description: Type of actor performing state transition

    # ==================== ORDER SCHEMAS ====================
    OrderCreate:
      type: object
      required:
        - customer_id
        - line_items
        - shipping_address
        - billing_address
        - payment_method
      properties:
        customer_id:
          type: string
          format: uuid
        line_items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/LineItemCreate"
        shipping_address:
          $ref: "#/components/schemas/Address"
        billing_address:
          $ref: "#/components/schemas/Address"
        payment_method:
          type: string
          example: "credit_card"
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        order_number:
          type: string
          example: "ORD-2024-001234"
        status:
          $ref: "#/components/schemas/OrderStatus"
        total_amount:
          type: number
          format: decimal
          example: 129.99
        currency:
          type: string
          example: "USD"
        payment_method:
          type: string
        payment_transaction_id:
          type: string
          nullable: true
        shipping_address:
          $ref: "#/components/schemas/Address"
        billing_address:
          $ref: "#/components/schemas/Address"
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/LineItemResponse"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        cancellation_reason:
          type: string
          nullable: true
    LineItemCreate:
      type: object
      required:
        - product_id
        - product_name
        - quantity
        - unit_price
      properties:
        product_id:
          type: string
          format: uuid
        product_name:
          type: string
          maxLength: 255
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: number
          format: decimal
          minimum: 0
    LineItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        product_name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
          format: decimal
        subtotal:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
    Address:
      type: object
      required:
        - street
        - city
        - state
        - zip
        - country
      properties:
        street:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 50
        zip:
          type: string
          maxLength: 20
        country:
          type: string
          maxLength: 50

    # ==================== RETURN SCHEMAS ====================
...
    StateTransition:
      type: object
      required:
        - new_state
      properties:
        new_state:
          type: string
        reason:
          type: string
          maxLength: 500
          nullable: true
    StateHistoryRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          type: string
          enum: [ORDER, RETURN]
        entity_id:
          type: string
          format: uuid
        previous_state:
          type: string
        new_state:
          type: string
        actor_id:
          type: string
          format: uuid
          nullable: true
        actor_type:
          $ref: "#/components/schemas/ActorType"
        trigger:
          type: string
          example: "API_CALL"
        metadata:
          type: object
          additionalProperties: true
        ip_address:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # ==================== COMMON SCHEMAS ====================
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_STATE_TRANSITION"
            message:
              type: string
              example: "Cannot transition from SHIPPED to PAID"
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "NOT_FOUND"
              message: "Order not found"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                field: "customer_id"
                issue: "Invalid UUID format"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []